Parses refactoring.guru problem domain description
