+++ crawler/scrapers/digital_lib_scraper/pipelines.py
