+++ crawler/scrapers/digital_lib_scraper/settings.py
