+++ crawler/scrapers/digital_lib_scraper/spiders/OldSpider/SpringframeworkSpider.py
